# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/github/kaiwinter/myatmo/main/MainActivity.java
#	app/src/main/res/values/strings.xml
#	build.gradle

+++ app/src/main/java/com/github/kaiwinter/myatmo/main/rest/model/DashboardData.java
@@ -6,13 +6,13 @@ import com.google.gson.annotations.SerializedName;
 public class DashboardData {
 
     @SerializedName("time_utc")
-    public Integer timeUtc;
+    public Long timeUtc;
     @SerializedName("Temperature")
     public Double temperature;
     @SerializedName("CO2")
-    public Integer cO2;
+    public Double cO2;
     @SerializedName("Humidity")
-    public Integer humidity;
+    public Double humidity;
     @SerializedName("Noise")
     public Integer noise;
     @SerializedName("Pressure")
